New implementation of BlockCachedFile
Instead of storing blocks sequentially in the file, store them in the
order they arrive.

This avoids long delays when accessing blocks e.g. at the end of files,
which previously required inserting padding for the middle part of the
file. The padding needs to be encrypted, which makes things slow.